refactor(ui): split history presenters
Extract shared history presenter helpers plus focused heatmap and detail presenter modules while keeping buildHistoryTableRowsViewModel as the stable top-level entrypoint.
